Release 0.7.0. See release notes or expand full commit message for details.
[Breaking]
Fixed WebRTC-Direct support added in Kubo 0.30.0.
Removed the Object API completely, since Kubo replaced it with the DAG API and no longer offers it.
Refactored the FileSystem API to:
  - Enable proper directory uploads via a new `AddAsync` method that takes File and Folder parts separately.
  - Bring `FileAddOptions` fully up-to-date with modern Kubo.
  - Remove `AddDirectoryAsync` (used missing Object API).
Updated several types to use one of int, long or ulong for Size matching the Kubo API.

[New]
Added IFilestoreApi and the corresponding types.
Added a new type `DagCid` that can be used to automatically create IPLD link references when serialized.
Added the IBlockStat type, analogous to the BlockStat struct in Kubo.

+++ src/DagLink.cs
@@ -14,7 +14,7 @@ namespace Ipfs
         /// <param name="name">The name associated with the linked node.</param>
         /// <param name="id">The <see cref="Cid"/> of the linked node.</param>
         /// <param name="size">The serialised size (in bytes) of the linked node.</param>
-        public DagLink(string? name, Cid id, long size)
+        public DagLink(string? name, Cid id, ulong size)
         {
             Name = name;
             Id = id;
@@ -68,7 +68,7 @@ namespace Ipfs
         public Cid Id { get; private set; }
 
         /// <inheritdoc />
-        public long Size { get; private set; }
+        public ulong Size { get; private set; }
 
         /// <summary>
         ///   Writes the binary representation of the link to the specified <see cref="Stream"/>.
@@ -100,20 +100,20 @@ namespace Ipfs
             }
 
             stream.WriteTag(3, WireFormat.WireType.Varint);
-            stream.WriteInt64(Size);
+            stream.WriteInt64((long)Size);
         }
 
-        private (string?, Cid, long) Read(Stream stream)
+        private (string?, Cid, ulong) Read(Stream stream)
         {
             using var cis = new CodedInputStream(stream, true);
             return Read(cis);
         }
 
-        private (string?, Cid, long) Read(CodedInputStream stream)
+        private (string?, Cid, ulong) Read(CodedInputStream stream)
         {
             string? name = null;
             Cid? id = null;
-            long size = 0;
+            ulong size = 0;
             while (!stream.IsAtEnd)
             {
                 var tag = stream.ReadTag();
@@ -126,7 +126,7 @@ namespace Ipfs
                         name = stream.ReadString();
                         break;
                     case 3:
-                        size = stream.ReadInt64();
+                        size = (ulong)stream.ReadInt64();
                         break;
                     default:
                         throw new InvalidDataException("Unknown field number");
@@ -153,6 +153,5 @@ namespace Ipfs
             Write(ms);
             return ms.ToArray();
         }
-
     }
 }
